requirements.txt
Flask==2.1.1
requests==2.32.3
selenium==4.3.0
python-telegram-bot==22.0
python-dotenv==0.19.0